17/12/24
Update the documentation.

On teams_frame place the button on a seperate frame within the team frame.

On left-top frame ,set the label of the database to show the number of searched players as well.

Bind the close window button to a save method that saves the current teams to a folder as a .pkl and .json files. If there are previous teams on that folder it deletes them.

Create the initialize_teams method that imports the existing teams on the folder to the Notebook as team frames.