17/12/24
On main: bind the tab selection of the notebook to the method on_tab_switch which updates the fields of the left-bottom frame to the values of the current tab. for convenience when there is a need to modify them.

On main and teams_frame: add also to the team information dictionary the Additional Info entry.