18/12/24
on main: fix a bug on initialize teams

on main: set the list of columns_to_show for the saved teams also in the json file of the teams info such that those columns will be displayed when open again.

on main: replace the yesno messeage window for saving upon closing with a yesnocancel window.

on teams_frame: modify the setting of the index column when export the team on a pdf or docx file, to avoid a warning.

on teams_frame: insert the pdf generation and docx code into try-except clause to catch the PermissionError when the same file is opened and cannot be replaced.

on table_frame and teams_frame: insert an if statement for the binding of the Treeview to the cursor to take the column name only when it is a valid sting and not an emtpy sting, which happens on empty regions of the Treeview that there are no columns, e.g. by shrinking the widths of the columns, releasing white space.