5/1/25
on main: fix a bug in the on_tab_switch_method; when there is no team left, clear the inputs to the left bottom fields.

on main: fix a bug on save_teams to save the teams even if no team is created, this will erase all saved teams.

on main: calculate the number of displayed rows in the table dynamically based on the avaliable screen height.